requirements.txt
geojson
python-dotenv
requests
Pillow